# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Oryon Gestao API - Documentacao de Endpoints",
        "description": "API RESTful profissional desenvolvida em Laravel 11 para gestao de departamentos, colaboradores e movimentacao estrategica de equipes.",
        "contact": {
            "name": "Jorge Carvalho",
            "email": "carvalhojorge7@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/departments": {
            "get": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Listar departamentos",
                "description": "Retorna listagem paginada de departamentos com suporte a filtro por status e cache inteligente.",
                "operationId": "e8478e3a557995e1da564073521762f6",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtro pelo status do departamento (ativo ou inativo)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ativo",
                                "inativo"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de registros por pagina",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numero da pagina",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listagem de departamentos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Criar departamento",
                "description": "Cadastra um novo departamento no sistema.",
                "operationId": "9a3a8ebf7326b72fd181c15603cdd21e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Engenharia de Software"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Equipe responsavel pelo desenvolvimento de software."
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "ativo",
                                        "enum": [
                                            "ativo",
                                            "inativo"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Departamento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validacao"
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/departments/{id}": {
            "get": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Exibir departamento e colaboradores",
                "description": "Retorna os dados de um departamento juntamente com seus colaboradores associados.",
                "operationId": "ba8d49f7da851631d9985375effc9b2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do departamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do departamento"
                    },
                    "404": {
                        "description": "Departamento nao encontrado"
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Atualizar departamento",
                "description": "Atualiza os dados de um departamento existente.",
                "operationId": "883e32516e81142e7441bad60554a993",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do departamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tecnologia e Inovacao"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descricao atualizada."
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "ativo",
                                        "enum": [
                                            "ativo",
                                            "inativo"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Departamento atualizado com sucesso"
                    },
                    "404": {
                        "description": "Departamento nao encontrado"
                    },
                    "422": {
                        "description": "Erro de validacao"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Remover departamento",
                "description": "Exclui um departamento sem colaboradores ativos vinculados.",
                "operationId": "4f695822af3d4a6db6f902305be6d0de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do departamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Departamento excluido com sucesso"
                    },
                    "422": {
                        "description": "Bloqueio por colaboradores ativos vinculados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "O departamento possui colaboradores ativos e nao pode ser excluido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Departamento nao encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/departments/{id}/transfer-employees": {
            "post": {
                "tags": [
                    "Transferencia"
                ],
                "summary": "Transferir colaboradores",
                "description": "Transfere todos os colaboradores ativos do departamento de origem para o departamento de destino com execucao atomica e rollback em caso de falha.",
                "operationId": "8108032ceed89d9b06721002eab92cdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do departamento de origem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "target_department_id"
                                ],
                                "properties": {
                                    "target_department_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transferencia concluida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Colaboradores transferidos com sucesso."
                                        },
                                        "transferred": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Origem e destino iguais ou erro de validacao"
                    },
                    "404": {
                        "description": "Departamento nao encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT retornado no endpoint /api/auth/login para autenticar as requisicoes.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}